requirements.txt
asgiref==3.8.1
Django==5.0.3
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.1
sqlparse==0.4.4
tzdata==2024.1
whitenoise==6.9.0
